Saving + Sectioned poi selection
Saving fully implemented
Poi's now generate in segments of the graph (segmented in 4 pieces, it will generate one poi in each segment before restarting). This gives even distribution of point location.

Added more things that are saved:
PoiInfo - graph the poi is in, poi location, and segment the poi was generated